% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practical_functions.r
\name{all_binom2num}
\alias{all_binom2num}
\title{From binomial to numerical}
\usage{
all_binom2num(data)
}
\arguments{
\item{dat}{dataset}
}
\value{
A new dataset. Convert every binomial column in the dataset as numerical to avoid errors
}
\description{
From binomial to numerical
}
