% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{bw_sens_pajd}
\alias{bw_sens_pajd}
\title{We want to create a table with different buffers for each variable in this case is adjusted by the number of covariates = bonferroni}
\usage{
bw_sens_pajd(dataset = data, forcing_var_name = "S",
  Y_name = "dropout", niter = 1000, covariates = c("sex",
  "HSHumanity", "HSTech", "HSOther", "hsgrade", "Y2004"),
  bandwidth = c(500, 1000, 1500), cut_value = 15000, whichunder = 1)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{forcing_var_name}{forcing_var_name}

\item{Y_name}{Y_name}

\item{niter}{niter}

\item{covariates}{covariates}

\item{bandwidth}{bandwidth}

\item{cut_value}{cut_value}

\item{forcing_bin_var_name}{forcing_bin_var_name}
}
\value{
data frame with variable and value and bandwidth
}
\description{
We want to create a table with different buffers for each variable in this case is adjusted by the number of covariates = bonferroni
}
