% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{da.bin.h02}
\alias{da.bin.h02}
\title{Impute missing compliance statuses under the NULL}
\usage{
da.bin.h02(theta, dat)
}
\arguments{
\item{theta}{theta}

\item{dat}{data set}
}
\value{
result.   ## Z=0 & W=0 (NT+C)
}
\description{
Impute missing compliance statuses under the NULL
}
