% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{da.gauss2}
\alias{da.gauss2}
\title{Calculate POSTERIOR MEAN - DA + MCMC}
\usage{
da.gauss2(theta, dat)
}
\arguments{
\item{theta}{theta}

\item{tol}{tol}
}
\value{
result. Impute missing compliance statuses under the null, ## Z=0 & W=0 (NT+C)
}
\description{
Calculate POSTERIOR MEAN - DA + MCMC
}
