% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{E.step.bin2}
\alias{E.step.bin2}
\title{Calculate MLE - EM}
\usage{
E.step.bin2(theta, dat)
}
\arguments{
\item{theta}{theta}

\item{dat}{dat}
}
\value{
result. We use Conditional probabilities using Bayes' Rule and we calculate the Expected compliance behavior
}
\description{
Calculate MLE - EM
}
