% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{EM.bin2}
\alias{EM.bin2}
\title{Calculate EM.bin}
\usage{
EM.bin2(dat, tol = 1e-04, maxit = 1000)
}
\arguments{
\item{tol}{tol}

\item{maxit}{maxit}

\item{G}{G}
}
\value{
result. tolerance fixed at 0.0001
}
\description{
Calculate EM.bin
}
