% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSided.R
\name{EM.gauss}
\alias{EM.gauss}
\title{Calculate EM.gaus}
\usage{
EM.gauss(dat, tol = 1e-04, maxit = 1000)
}
\arguments{
\item{tol}{tol}

\item{maxit}{maxit}

\item{G}{G}
}
\value{
result. tolerance fixed at 0.0001
}
\description{
Calculate EM.gaus
}
