% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{fisher_exact_value}
\alias{fisher_exact_value}
\title{Calculating Fisher Exact p-value}
\usage{
fisher_exact_value(dataset, forcing_bin_var_name = "Z",
  Y_name = "dropout", niter = 100)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{forcing_bin_var_name}{forcing_bin_var_name}

\item{Y_name}{Y_name}

\item{niter}{niter}
}
\value{
data frame with variable and value
}
\description{
Calculating Fisher Exact p-value
}
