% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{fuzzy_fep2sided}
\alias{fuzzy_fep2sided}
\title{Fuzzy - FEP binary two sided}
\usage{
fuzzy_fep2sided(dataset, Y, W, Z, Y_name, M2 = 10)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{W}{selected}

\item{Y_name}{Y_name}

\item{M2}{number of iterations}

\item{forcing_bin_var_name}{forcing_bin_var_name}

\item{forcing_var_name}{forcing_var_name}

\item{niter}{niter}

\item{bandwidth}{bandwidth}

\item{cut_value}{cut_value}
}
\value{
data frame with variable and value and bandwidth
}
\description{
Fuzzy - FEP binary two sided
}
