% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{fuzzy_neyman}
\alias{fuzzy_neyman}
\title{Fuzzy - neyman}
\usage{
fuzzy_neyman(Y, Wh, Z, cin = 95)
}
\arguments{
\item{Y}{Outcome variable}

\item{Wh}{selected records}

\item{Z}{forcing binary variable}
}
\value{
data frame with variable and value and bandwidth
}
\description{
Fuzzy - neyman
}
