% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSided.R
\name{mcmc.bin.h0}
\alias{mcmc.bin.h0}
\title{Calculate mcmc.bin.h0}
\usage{
mcmc.bin.h0(n.iter, n.burn, dat)
}
\arguments{
\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}
}
\value{
result.   ## Z=0 & W=0 (NT+C)
}
\description{
Calculate mcmc.bin.h0
}
