% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{mcmc.bin2}
\alias{mcmc.bin2}
\title{Calculate mcmc.bin}
\usage{
mcmc.bin2(n.iter, n.burn, dat)
}
\arguments{
\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}
}
\value{
result.
}
\description{
Calculate mcmc.bin
}
