% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{mcmc.gauss.h02}
\alias{mcmc.gauss.h02}
\title{Calculate mcmc.gaus.h02}
\usage{
mcmc.gauss.h02(n.iter, n.burn, dat)
}
\arguments{
\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}
}
\value{
result.   ## Z=0 & W=0 (NT+C)
}
\description{
Calculate mcmc.gaus.h02
}
