% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSided.R, R/TwoSided.R
\name{mcmc.gauss}
\alias{mcmc.gauss}
\title{Calculate mcmc.gaus}
\usage{
mcmc.gauss(n.iter, n.burn, dat)

mcmc.gauss(n.iter, n.burn, dat)
}
\arguments{
\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}

\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}
}
\value{
result.

result.
}
\description{
Calculate mcmc.gaus

Calculate mcmc.gaus
}
