% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSided.R
\name{mcmc.gauss2}
\alias{mcmc.gauss2}
\title{Calculate mcmc.gaus}
\usage{
mcmc.gauss2(n.iter, n.burn, dat)
}
\arguments{
\item{n.iter}{number of iterations}

\item{n.burn}{burn-in step}

\item{dat}{data set}
}
\value{
result.
}
\description{
Calculate mcmc.gaus
}
