% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{rd_filtering}
\alias{rd_filtering}
\title{Filtering the datasaet with ranges}
\usage{
rd_filtering(dataset = last, covariate = "AGE", num_out = 4,
  CT = "W", Nsim = 1000, cutvar = 30, low_lim = NULL,
  high_lim = NULL)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{covariate}{One or more covariate}

\item{CT}{control and treatment variable}

\item{Nsim}{Number of simulations}

\item{cutvar}{number to divide the treatment}

\item{low_lim}{lowest limit of each of the covariate}

\item{high_lim}{lowest limit of each of the covariate}
}
\value{
pvalues and other properties
}
\description{
Filtering the datasaet with ranges
}
