% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{rd_pvalue}
\alias{rd_pvalue}
\title{Calculate de regression discontinuity p-value}
\usage{
rd_pvalue(dataset = last, covariate = "AGE", CT = vec, Nsim = 1000,
  cutvar = 30, low_lim = NULL, high_lim = NULL)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{covariate}{One or more covariate}

\item{CT}{control and treatment variable}

\item{Nsim}{Number of simulations}

\item{cutvar}{number to divide the treatment}

\item{low_lim}{lowest limit of each of the covariate}

\item{high_lim}{lowest limit of each of the covariate}
}
\value{
pvalues and other properties
}
\description{
Calculate de regression discontinuity p-value
}
