% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{sharp_fep}
\alias{sharp_fep}
\title{Calculating Fisher Exact p-value}
\usage{
sharp_fep(dataset, forcing_bin_var_name = "Z", Y_name = "dropout",
  niter = 100)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{forcing_bin_var_name}{forcing_bin_var_name}

\item{Y_name}{Y_name}

\item{niter}{niter}
}
\value{
data frame with variable and value
}
\description{
Calculating Fisher Exact p-value
}
