% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{sharp_neyman}
\alias{sharp_neyman}
\title{SHARP RDD: RANDOMIZATION-BASED INFERENCE - NEYMAN APPROACH}
\usage{
sharp_neyman(Y, Z, cin)
}
\arguments{
\item{Y}{Outcome variable}

\item{Z}{forcing binary variable}

\item{cin}{Confidence interval}
}
\value{
a vector with values
}
\description{
SHARP RDD: RANDOMIZATION-BASED INFERENCE - NEYMAN APPROACH
}
