% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{sharp_neyman_bw}
\alias{sharp_neyman_bw}
\title{Sharp - neyman bandwidth}
\usage{
sharp_neyman_bw(dataset = data, forcing_var_name = "S",
  Y_name = "dropout", niter = 1000, bandwidth = c(500, 1000, 1500),
  cut_value = 15000, whichunder = 1, cin = 95)
}
\arguments{
\item{dataset}{The dataset with the variables}

\item{forcing_var_name}{forcing_var_name}

\item{Y_name}{Y_name}

\item{niter}{niter}

\item{bandwidth}{bandwidth}

\item{cut_value}{cut_value}

\item{forcing_bin_var_name}{forcing_bin_var_name}
}
\value{
data frame with variable and value and bandwidth
}
\description{
Sharp - neyman bandwidth
}
