% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDis_function.r
\name{Tave}
\alias{Tave}
\title{Tave}
\usage{
Tave(x, z)
}
\arguments{
\item{x}{The dataset with the variables}

\item{z}{forcing_bin_var_name}
}
\value{
data frame with variable and value and bandwidth
}
\description{
Tave
}
