% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practical_functions.r
\name{which.format}
\alias{which.format}
\title{Detect the format of the file}
\usage{
which.format(pathfile)
}
\arguments{
\item{pathfile}{a string with the name and format of the file}
}
\value{
the format of the file
}
\description{
Detect the format of the file
}
